clear error on other chat click
