+++ app/src/main/java/org/ole/planet/myplanet/ui/chat/ChatDetailFragment.kt
@@ -4,7 +4,6 @@ import android.os.Bundle
 import android.text.Editable
 import android.text.TextUtils
 import android.text.TextWatcher
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -64,13 +63,11 @@ class ChatDetailFragment : Fragment() {
                 val message = "${fragmentChatDetailBinding.editGchatMessage.text}".replace("\n", " ")
                 mAdapter.addQuery(message)
                 if(_id != ""){
-                    Log.d("continue","chat");
                     val continueChatData = ContinueChatModel(data = Data(message, _id, _rev), save = true)
                     val jsonContent = Gson().toJson(continueChatData)
                     val requestBody = RequestBody.create(MediaType.parse("application/json"), jsonContent)
                     continueChatRequest(requestBody)
                 } else {
-                    Log.d("new","chat");
                     val chatData = ChatRequestModel(data = ContentData(message), save = true)
                     val jsonContent = Gson().toJson(chatData)
                     val requestBody = RequestBody.create(MediaType.parse("application/json"), jsonContent)
@@ -93,6 +90,8 @@ class ChatDetailFragment : Fragment() {
 
         sharedViewModel.getSelectedChatHistory().observe(viewLifecycleOwner) { conversations ->
             mAdapter.clearData()
+            fragmentChatDetailBinding.editGchatMessage.text.clear()
+            fragmentChatDetailBinding.textGchatIndicator.visibility = View.GONE
             for (conversation in conversations) {
                 val query = conversation.query
                 val response = conversation.response
@@ -196,7 +195,6 @@ class ChatDetailFragment : Fragment() {
     }
 
     fun clearChatDetail() {
-        Log.d("called", "called")
         mAdapter.clearData()
         _id = ""
         _rev = ""
