--- app/src/main/java/org/ole/planet/myplanet/ui/chat/ChatHistoryItemClickListener.kt
@@ -1,8 +1,0 @@
-package org.ole.planet.myplanet.ui.chat
-
-import io.realm.RealmList
-import org.ole.planet.myplanet.model.Conversation
-
-interface ChatHistoryItemClickListener {
-    fun onChatHistoryItemClicked(conversations:RealmList<Conversation>, _id: String, _rev: String)
-}
