+++ app/src/main/java/org/ole/planet/myplanet/ui/chat/ChatRequestModel.kt
@@ -10,3 +10,14 @@ data class ChatRequestModel(
 data class ContentData(
     @SerializedName("content") val content: String
 )
+
+data class ContinueChatModel(
+    @SerializedName("data") val data: Data,
+    @SerializedName("save") val save: Boolean
+)
+
+data class Data(
+    @SerializedName("content") var content: String,
+    @SerializedName("_id") var Id: String,
+    @SerializedName("_rev") var rev: String,
+)
